разбил на классы, сделал run task в gradle
