+++ .idea/gradle.xml
@@ -24,6 +24,7 @@
             <option value="$PROJECT_DIR$/L26-di/class-demo" />
             <option value="$PROJECT_DIR$/L26-di/homework" />
             <option value="$PROJECT_DIR$/L28-springDataJdbc" />
+            <option value="$PROJECT_DIR$/L32-executors" />
             <option value="$PROJECT_DIR$/buildSrc" />
             <option value="$PROJECT_DIR$/hw19-jdbc" />
             <option value="$PROJECT_DIR$/hw19-jdbc/demo" />
