+++ L32-executors/src/main/java/com/github/vladimirplotnikov/homework/ThreadMode.java
@@ -1,0 +1,5 @@
+package com.github.vladimirplotnikov.homework;
+
+public enum ThreadMode {
+    Counter, Printer
+}
